Fix dtype mismatches in STT img_projector and training_losses
Two additional float32/bfloat16 crashes:
- stt.py forward(): img_projector received float32 latents (from VAE
  encode_to_z) against bfloat16 weights; added .to(target_dtype) cast.
- stt.py evaluate(): same issue in the eval path.

Also resolve silent mixed-precision arithmetic in training_losses:
cast target and x_t to pred.dtype (bfloat16) after the DiT forward so
all operands in the MSE and predict computations share a single dtype.

+++ models/flux_dit.py
@@ -145,7 +145,9 @@ class FluxDiT(nn.Module):
         target = img - noise
         pred = self(img=x_t, img_ids=img_ids, cond=cond, cond_ids=cond_ids, timesteps=t.reshape(-1), y=y, guidance=guidance)
         assert pred.shape == target.shape == img.shape
-        predict = x_t + pred * (1. - t)
+        target = target.to(pred.dtype)
+        x_t = x_t.to(pred.dtype)
+        predict = x_t + pred * (1. - t.to(pred.dtype))
         terms["mse"] = mean_flat((target - pred) ** 2)
         
         terms["loss"] = terms["mse"].mean()
